Create abstract `StackExchangeAccount` class
